Проработка чата
