+++ app/src/main/java/com/example/appshops/main/fragments/FragmentMessageScreen.kt
@@ -7,16 +7,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.appshops.R
+import com.example.appshops.authorization.viewmodel.MainViewModel
 import com.example.appshops.main.adapter.AdapterRecyler
 import com.example.appshops.model.User
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.database.DatabaseError
-import com.google.firebase.database.ValueEventListener
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
 
 
 class FragmentMessageScreen : Fragment() {
@@ -24,7 +21,7 @@ class FragmentMessageScreen : Fragment() {
 
     private lateinit var recylerView: RecyclerView
     private lateinit var adapterMassage: AdapterRecyler
-
+    lateinit var viewModel: MainViewModel
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -47,43 +44,21 @@ class FragmentMessageScreen : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        viewModel = ViewModelProvider(this).get(MainViewModel::class.java)
         initViews(view)
         adapterMassage = AdapterRecyler()
         recylerView.layoutManager = LinearLayoutManager(context)
         recylerView.adapter = adapterMassage
-        databaseRead()
+        viewModel.databaseRead()
         adapterMassage.setOnClickItemListener(object : AdapterRecyler.OnClicklistener {
             override fun onUserClickIte(user: User) {
                 Log.d("MessageScreen", "Onbclick")
             }
-
         })
     }
 
     override fun onDestroy() {
         super.onDestroy()
-    }
-
-    fun databaseRead() {
-        val database = Firebase.database
-        val myRef = database.getReference("user")
-        myRef.addValueEventListener(object : ValueEventListener {
-            override fun onDataChange(snapshot: DataSnapshot) {
-                try {
-                    for (dataSnapshot in snapshot.children) {
-
-                    }
-                } catch (error: Exception) {
-                    Log.d("FirebaseValue", "Value is: $error")
-                }
-
-            }
-
-            override fun onCancelled(error: DatabaseError) {
-                Log.w("ErrorFirebase", "Ошибка", error.toException())
-            }
-
-        })
     }
 
     fun initViews(view: View) {
